+++ packages/dialect-react-sdk/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - feature: introduce `useNotificationSubscriptions` hook to configure notifications
 - feature: introduce `useDappNotificationSubscriptions` hook to fetch notification subscriptions on behalf of dapp
+- fix: types for react 18
 
 ## [1.0.0-beta.15] - 2022-07-25
 - fix: notification togglers